feat: Add format functions
