+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './cacheable';
 export * from './sleep';
 export * from './env';
+export * from './format';
