+++ src/view/BlocksView.tsx
@@ -1,8 +1,8 @@
 'use client';
 
-import pluralize from 'pluralize';
 import { Table, Cell } from 'gexii/table';
 import { Button, Pagination, Stack } from '@mui/material';
+import { formatNumber, timeAgo } from 'src/utils';
 import Link from 'next/link';
 import { useRouter } from 'next/navigation';
 import type { PreviewBlockDto } from 'src/services';
@@ -97,28 +97,4 @@ export default function BlocksView({ blocks, page, slot, now }: BlocksViewProps)
       </Stack>
     </Stack>
   );
-}
-
-const intervals = {
-  second: 1000,
-  minute: 60 * 1000,
-  hour: 60 * 60 * 1000,
-  day: 24 * 60 * 60 * 1000,
-  week: 7 * 24 * 60 * 60 * 1000,
-  month: 30 * 24 * 60 * 60 * 1000,
-};
-function timeAgo(diff: number) {
-  if (diff < intervals.second) return 'Just now';
-  const hitInterval = Object.entries(intervals).findLast(([_, value]) => diff >= value);
-
-  // [NOTE] Error prevention, this should never happen
-  if (!hitInterval) return 'Unknown time';
-
-  const unit = hitInterval[0];
-  const value = Math.floor(diff / Number(hitInterval[1]));
-  return `${value} ${pluralize(unit, value)} ago`;
-}
-
-function formatNumber(num: number | bigint) {
-  return num.toString().replace(/\B(?<!\.\d*)(?=(\d{3})+(?!\d))/g, ',');
 }
